Ability to teardown a continuation context with a returned function.
When you yield a function, that function gets passed the
execution. Previously, the only way to register teardown was by adding
a `finally()` handler onto that execution. However, if you have 5
steps in a process, none of the them will have their teardowns run
until the full process exists. That could be a very long time.

This lets the control function return a "release" function that is
invoked as soon the context resumes. So, for example, the timeout
helper returns a function that removes the timeout.

```js
export function timeout(durationMillis = 0) {
  return function(execution) {
    let timeoutId = setTimeout(() => execution.resume(), durationMillis);
    return () => clearTimeout(timeoutId);
  };
}
```