Remove Task nomenclature for now.
Prior to this point, we'd been using "Task" as a stand-in for just a
generator function. But this is a high value term that we don't want
to load too soon.

Let's be conservatice about this and call generators just "procs" and
make what was currently Task, but was really just a way of creating a
generator function out of anything, and just call it that.