Capability to yield to promises.
Structured concurrency lets you manage asynchronous process and tear
them down and set them up as transactional units. However, the primary
mechanism for most async apis is the promise. We need a way to be able
to easily yield an async execution until the promise either completes,
or raise an error when it rejects.

This adapts the promise api using an execution controller named
`promiseOf` which resumes an execution when the promise resolves and
throws an error inside the execution when the promise rejects. Using
this basic primitive, we can adapt most apis to an async execution.