Implicitly treat promises as an execution step.
The most common form of managing asynchrony in JavaScript is the
promise. That's why we included the `promiseOf` helper to [adapt a
promise into an execution context][1].

However, after using the api to build something, it became immediately
apparent that the most intuitive thing was to be able to yield to any
promise, and that it was pretty much busy work to have to wrap a
promise in a call to `promise-of` yourself.

This change makes wrapping in `promise-of` implicit. in other words,
if you yield to a promise, it will automatically create a promise-of
execution controller to resume evaluation once the promise has
settled.

For example:

```js
function getJSON(url, init) {
  return function*() {
    let response = yield promiseOf(fetch(url, init));
    return yield promiseOf(response.json());
  }
}
```

becomes:

```js
function getJSON(url, init) {
  return function*() {
    let response = yield fetch(url, init);
    return yield response.json();
  }
}
```

[1]: https://github.com/cowboyd/effection.js/pull/4