Add Typescript typings
Typescript provides a very nice level of auto-documenation and
discoverability to its users, especially for a strange new api like
the one provided by `effection`. By definining types for effection
functions, this will let us bring it easily to the typescript
community and allow it to clear the bar set by that community.

Rather than implement types at the `DefinitelyTyped` repository where
they could conceivably get out of sync with this package, instead this
adds an `index.d.ts` directly to this project that will ship with
every NPM package.

To ensure that the typescript typings work, the typescript.test.js
loads a set of files and then asserts that they either typecheck or
that they dont.