Upgrade `nyc` coverage tool
We want to be above suspicion when it comes to security, and as such,
there were a number of package with vulnerabilities that were being
pulled in via the coverage tool. Technically, this shouldn't be a
problem since coverage is only happening when running tests and not
when effection is being used on the server, but we'll address it
anyway.

Upgrade `nyc` so that transitive dependencies are secure.