Make `fork()` a simple function
Asynchronous operations are executed with the `fork()` method on
`Execution` which is used to mark which exact process that you're
forking off of. This is necessary because where you fork from
determines who your parent will be and what happens when that parent
is halted, or when you finish or error out who you need to report to.

The way we were accessing the current execution was via the `this`
parameter inside of our operation functions. It was doen this way
because it was easy. If you have a direct reference to the current
execution, then it is very easy to fork off operations from
it. Another rationale was that the `fork()` method _is_ effectful and
so better to mark it as special by associating it with the `this`
parameter. That way, it can be clearly seen as an imperative API.

It hasn't really panned out that way. In this day and age, it's
is a very odd paradigm to be following as use of `this` is
shunned. It also is a point of strangeness with the rest of the API
since it is otherwise very functional in nature.

Since we can only have one execution running at a time (JS is single
threaded after all), this change tracks a "current" execution which is
used as the execution context any time that `fork` is called.

before:

```ts
export default function* main(env: Env): Sequence {
  this.fork(webAPI, [env.port]);
  this.fork(rtmAPI, [env.slackOAuthToken]);
}
```

after:

```ts
import { fork } from 'effection';

export default function* main(env: Env): Sequence {
  fork(webAPI, [env.port]);
  fork(rtmAPI, [env.slackOAuthToken]);
}
```