Make generators (not just generator functions) a valid operation.
Previously, the only way to define an operation with a co-routine was
by passing a generator function to `yield`. That meant we had to do a
lot of silly argument passing around the codebase just to support
passing arguments to functions. Like:

```js

function* add(left, right) {
  return left + right
}

function* top() {
  let one = yield function*() { return 1; }
  let two = yield function*() { return 2; }
  return yield call(add, one, two);
}
```

But JavaScript supports passing arguments to functions just fine,
right? Wouldn't it be better if we could just do it like this?

```js
function* add(left, right) {
  return left + right
}

function* top() {
  let one = yield function*() { return 1; }
  let two = yield function*() { return 2; }
  return yield add(one, two);
}
```

That's what this change does. By allowing you to invoke the generator
function directly, we can do away with the `call` function altogether.