Unify execute() and fork()
With `fork()` now being a static function, and operations all being
atoms that don't have arguments, the signature for `fork()` and
`execute()` became exactly the same:

```ts
fork(operation) => Execution
execute(operation) => Execution
```

It wasn't hard to see that `execute` was nothing more than a "root"
fork. Or a just fork that had no existing context, so it made it much
more conceptually simple to eliminate one in favor of the other.

This removes the `execute` function altogether, and adds a base case
to `fork()` where if the current execution is not defined, it creates
a brand new execution to serve as the root context.