➖Remove unused code
The `Gate` object was an artifact of the original concept of monadic
effects. While the underlying concepts are the same, the fluent,
generator based "vanilla JS" approach makes it completely
necessary. Also removed is the `generator.test` which had several
experiments with generators to better understand how they work.

This removes all of those since they have been supplanted by the
current API.