🐛Don't mark a running fork as completed when a child is halted
When you have the case of a parent A that is running, and also that
parent has a forked child `B` that becomes halted. A, because it is
still yielding should continue running. If on the other hand, `A` is
waiting, then it should be completed because that was the last
blocking child.

```
A
 - B
```

Howewer, we only checked that the parent did not have blocking
children, and not that it was also just waiting (and not running).

This adds the check to the `join()` method to ensure that running
parents continue running even when a child is halted.