🐛Always cleanup children
When debugging an effection application, we noticed that in the case
where you had halted children, they were not being removed from
parents that were still running. That meant that you have a fork tree
looking like:

```
Fork {
   state: waiting
   operation: main
   children:
     Fork {
       state: running
       operation:
       children:
     }
     Fork {
       state: running
       operation:
       children:
           Fork {
             state: halted
             operation: setupConnection
             children:
           }
     }
}
```

Which means that if the the parent is long running, then you can
accrue lots of halted children which is a memory leak.

This makes sure that _every_ time that a child joins its parent, it is
removed from the list of children. This is safe because join _only_
happens on finalization of a child.