Add `halt()` to execution signature
For each execution context, there are three possible outcomes:
completed, errored, or halted. The execution interface however only
accounts for two of them in the official type signatures.

Because cancellation through halt is a critical api, and is already in
use in other projects, this adds it to the official type declarations.