Add `id` field to each execution
It's nice when debugging to see how forks relate to each other both
inside and outside of code. To do this you need a consistent way to
identify a fork.

This adds an `id` field which is allocated from a static id sequence
inside the Fork constructor. I've opted to install the class
properties syntax since it is so pervasive and cleaner than using a
module variable.

I also like the idea of being able to read it as a property from the
`Fork` constructor rather than have it completely hidden inside a
module variable.

resolves #27