Run type tests in separate suite
The TypeScript tests are slow because they need to invoke the
TypeScript compiler on each test which is actually quite slow. We
don't want to run them with the default `yarn test` command, but we do
want to run them on CI so that broken type declarations still can't
get through.

This change introduces a new npm script, `test:types` which only runs
the `tests/typescript.js` suite, while the normal `yarn test` will
only run those files ending in `.test.js`.

The Circle CI job has been refactored to take a parameter of which
yarn script to run after the node_modules have been installed.  Then,
a second workflow is introduced that runs not just the tests, but also
the type tests.