Add monitor function which doesn't block parent from shutting down
This comes up time and again, where we want to have a fork, but do not want the fork to block the parent from shutting down. Where this is common is in the error monitor pattern that has emerged in bigtest-server. Hence the function name.

So for example:

```
  let errorMonitor = fork(function*() {
    let [error]: [Error] = yield on(child, "error");
    throw error;
  });
```

This is tricky because the error monitor will block the fork from shutting down. This is likely not what we want. Instead we want the monitor to shut down if the parent exits. To accomplish this today, we need to wrap the entire rest of the function in a try-finally.

```
  let errorMonitor = fork(function*() {
    let [error]: [Error] = yield on(child, "error");
    throw error;
  });

  try {
    ...
    yield
    ...
  } finally {
    errorMonitor.halt();
  }
```

This is both ugly and error prone. With the monitor function we can write this more naturally:

```
  monitor(function*() {
    let [error]: [Error] = yield on(child, "error");
    throw error;
  });
```