🐛Fix generator handling of caught errors
One of the niceties of the generator operations is that they can
elegantly express error handling in a computation by using the native
try/catch syntax.

```js
for (let tries = 0; tries < 5; tries++) {
  try {
    return yield dangerousOperation;
  } catch (error) {
    continue;
  }
}
throw new Error('Max retries exceeded. Giving up!');
```

In effection, if `dangerousOperation` fails, then we need to throw
that error inside the generator. If the generator doesn't catch that
error, then `generator.throw(error)` re-throws that error. We were
handling that case and failing the generator operation as a whole.

However, if the generator _catches_ the error, as in the example
above, then when we call  `generator.throw(errror)` it will return the
next iteration value. We were calling resume with the iterator value,
and not the next operation. That meant that you only saw weird
behavior when you caught an error and then re-tried the error handling
block.

This abstracts the handling of each iterator value with the `advance`
function, and makes sure that whether we are continuing a generator
because of a failure or because of a resumption of the computation, we
still handle the operation uniformly.