Use more abstract typing
As we use effection more, it's clear that it isn't really helpful when
it comes to composability to think of anything other than
`Operation`s. It actually adds complexity to think of a generator
function as anything other than just another function that happens to
return an operation.

This lets you compose normal functions with generators. So, for
example, the callback for a childProcess supervisor takes a function
that accepts a child and returns another operation:

```ts
type UseChild = (child: ChildProcess) => Operation;

let getPid: UseChild = child => ({ resume }) => resume(child.pid);

let awaitExit: UseChild = function*(child) {
  yield on(child, 'exit');
}
```

One uses a generator, but the other is a controller function, but
their both assignable to the abstract Operation constructor. It's
not helpful to distinguish between or declare them as their concrete
types, and in fact, it's probably harmful in that it interferes with
the mental model.

This change normalizes against `Operation` as the fundamental unit of
operation, and actually un-exports the concrete types. Where as
before, you would declare your generator function to return a
sequence, now it returns an Operation.

before

```ts
function* doSomething(arg: Arg): Sequence {

}
```

after

```ts
function* doSometing(arg: Arg): Operation {

}
```

This help re-enforce the idea that a generator is just another way to
define a function that returns an operation.