Remove implicit mailboxes
Stateful mailboxes are a powerful primitive in asynchronous
programming. However, it turns out that these are pretty
straightfoward to implement as an addon package to Effection, and
as such, do not need to be in effection core.

This removes all mailbox-related code. It's expected that mailboxes
will still have official support from effection, but that they will be
part of the `@effection/events` package that has utilities for working
with events and event emitters.