Catch errors thrown in ensure block
Closes #78

If an error ever occurrs in an ensure block, then that is a serious
problem. We should inform the user that this is not good by printing a
huge nasty warning to the console.

However, if such a situation does occur, at least it should not leave
the process tree in an inconsistent state. Other exit hooks should still
run, and other processes should still shut down. Currently if such an
error does occur, it will blow the stack, and cause finalization to stop
abrubtly. This is very much not good.

Here we're rescuing any such errors, printing them to the console, but
crucially we are NOT re-raising them. This allows us to proceed with
finalization, which will leave the process tree in a more consistent
state.

It's worth noting though that such a situation really is very
undesirable. The Rust language does a hard abort of the process with an
error exit code if a destructor panics, we cannot do something similar
in effection, since we want to run in browsers, and browsers cannot
hard-exit.