re-jigger how types are built
We never did diagnose why this was necessary, but rearranging it into
this configuration did work.

What we did was to switch the symlinking from `index.d.ts` <-
`types/index.d.ts` to be the other way around:

`index.d.ts` -> `types/index.d.ts`

`dtslint` was happy with that 🤷🏼

But that made `@pika` unhappy, so we had to switch to the specific
typescript plugin. Hopefully this works.