Extract the much used `main` method into its own package.
Motivation
----------

When you're managing a node process, there are things that you almost
always want to do: setup a SIGINT handler to interrupt the process,
catch errors and log them to the console, and exit when appropriate
with the proper code. The server package had a very handy main method
that was used in several places via copy paste.

Approach
--------

This consolidates all of those main methods down into a single
location that can be shared by all.