Use ExecutionContext subclass for generators
The first implementation using an unassociated ExecutionContext turned out to be problematic, because forks created within such a child ended up without a parent. This implementation instead creates a hidden parent context for all generator operations, which are then children of this context. We are overriding `trapExit` to customize what happens when a child is finished. This also means we can get rid of the odd continuation bit in ExecutionContext.

This makes top level forks work as intended.

There is a bit of a weird workaround to make `fork` implementable as a regular operation:

If a required context is returned, it will NOT block until completed, instead it will be passed to the parent in its current state. I think this is reasonably intuitive. The weirdness comes in that if a required context is returned from the root context, then it WILL block. I think in practice this is probably fine, but it is a bit odd.