Add `Deferred` an alternative to explicit control functions
One of the ongoing discussions we've been having is that we want to
move away from control functions and really focus on having generators
and promises be the base primitives that people work with in
effection.

As a baby step towards that goal, this introduces a `Deferred` function
that returns an object containing a `resolve` / `reject` / `promise`
calling `resolve()` will resolve `promise`, and calling `reject`, will
reject `promise`.

This provides a simple control pattern very similar to the exsting
control functions, but based on standard and familiar apis: promise
and generator.

> Ok, generators may not feel familiar, but they are at least
> "familiar-ish" in the sense that they're just JS code.

Thus, we can re-write any control function to be `Deferred` +
`try/finally` with a yield to the futures promise.

As an example, `timeout` is written in this way.