Don't crash when an operation returns a null value
We need to check for resources on return values of operations in order
to link them to the parent context. To do this, we check that the
object can receive property access. However, `typeof null` is actually
`"object"`.

This adjusts the check to cover both undefined and null (as well as
any other object type)