Add event package for working with events
Working with event emitters has been a really common need when working with Effection. Over time we evolved some helpers for making this more efficient.

We have also extensively used mailboxes for working with events. However, with the advent of resources, it is now pretty easy to implement and efficient and easy to use event iterator. I've called this operation `on`. This is a more light-weight solution than mailboxes. There is some overlap in functionality, but I think this is acceptable. Basically we could replace a lot of the mailboxes in bigtest with this, and it would make it clearer where we actually need mailboxes and where we just use them to subscribe to events.