Use yarn for prepacking and publishing
We wer using yarn for install and npm for publishing. Also, now that
we have the node package which will depend on effection events, it's
important that all interfaces are built in dependency order according
to the root package.json

This normalizes and upgrades the release and preview workflows, while
adding a before_all hook to both to run a prepack on the entire repo.