Make resolution of `once` asynchronous
If we run the `resume` in the same tick as the listener, then we can run into a problem with the event listener being removed in the same tick as it is running. Some badly behaved event libraries (looking at you yaeti) assume that this will not happen 🙄

We really shouldn't have to do this, but here we are.