Respond to SIGTERM in @effection/node/main
We were only responding to `SIGINT` inside our `main` handler. What
this meant, was that if an effection process contained sub-effection
processes, managed by one of the `child_process` resources, then
because the main context was not getting torn down, the subsequent
resources that represented child process were not getting sent their
shutdown signals.

This adds SIGTERM to the list of signals that will cause a shutdown of
the main effection context. That way, effection processes that manage
other process will shut them down whenever they are themselves
terminated.

[1]: https://www.gnu.org/software/libc/manual/html_node/Termination-Signals.html