Implement subscriptions level 2
According to the RFC, the second level after attaining the core
subscription operation `createSubscription` is to have a way to

1. Make any object subscribeable
2. Given any subscription, have a way to transform that subscription
by mapping or filtering.

Approach
--------

Define an "interop" symbol to be exported that should return a
subscribe operation:

```ts
import { SymbolSubscribeable, Subscribeable } from '@effection/subscription';
import { Message } from 'socket';

class WebSocket implements Subscribeable<Message, void> {
  constructor(public socket: Socket) {}

  [SymbolSubscribeable]() {
    return on(this.socket, 'message');
  }
}
```

Now you can iterate over the array using the builtin `forEach`
function:

```ts
let ws = new WebSocket(socket);
yield forEach(ws, function*(message) {
  console.log('received message');
});
```

A operation that produces a subcription, or a subscribeable can be
transformed with higher-order functions using the `Subscribeable.from`
function.

```ts
yield Subscribeable.from(ws)
  .map(message => JSON.parse(message))
  .filter(message => message.type == 'command')
  .forEach(function*(command) {
    yield handleCommand(command);
  })
```

Todos & Open Questions
----------------------

There's a bit of an asymmetry in that the "return value" of a
subscription is distinct from the members of the individual items of
the subscription. Operations like map, and filter _do not_ operate on
the return value, only the individual items.

Learning
--------

Apparently, you can use [`Symbol.for()`][1] to lookup/create a symbol in a
runtime-wide fashion. We use that instead of the normal `Symbol()`
constructor for the exported interop delegate. This way, no matter
which "version" of the `@effection/subscription` module you are
getting the  `SymbolSubscribeable` from, it will return the same
Symbol reference.

[1]: https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/Symbol/for