Only require the event listener methods we use
In [BigTest][1], we're settling on using the `ws` websocket library
across the board, but sockets from that library don't implement the
full `EventTarget` api, only `{add,remove}EventListener`. This means
that we can't use those sockets easily because they cannot be passed
to the underlying subscription mechanism which is `on()` from
'@effection/events'.

This relaxes the requirement of what can constitute an EventSource to
just things that have the `addEventListener` and `removeEventListener`
methods, since those are the only ones we use.

[1]: https://github.com/thefrontside/bigtest