Use a single, unique symbol for resourcels
When installing bigtest into an existing codebase, we ended up with
multiple versions of effection. That meant that resources created with
one version were not being recognized by another. An example was when
we created a ParceProcess to build the manifest. Because the calling
context (@bigtest/server) did not recognize the child resource that
was returned by @bigtest/parcel as a resource. As a result, the child
process was shutdown immediately.

This uses the [Symbol.for][1] method to always return the same symbol
no matter which version of effection that it is imported from.

Note that this will only work for all versions of effection that
include this specific change. Previous versions will continue to have
the same problem.