Convert effection into normal dependency
Whenever there is a minor change to a peer dependency, this causes
changesets to want to release a new major version of the
dependent, which isn't what we really want.

Now that we've made effection have a universal unique symbol, we
should theoretically be allowed to have several versions of the
library that are in-range kicking around in the runtime.

In the interest of keeping things simple, this converts all peer
dependencies on effection into regular dependencies. We can always
convert back to peer dependencies in the event that problems do
arise.