Convert fetch response into resource
We were running into an issue where if the bytes of the body were not
fully downloaded by the time headers had been received, the request
was being aborted, which was causing the `response.text()` promise to
become rejected.

This converts the `Response` object to a resource so that only when
the response is no longer used will the request be aborted.

There were a bit of gymnastics that we had to adopt in order to make
this work since it is possible that fetch could be halted _before_ the
underlying response was returned from native fetch. In that event, we
still need to abort the request.